share deaduntils between hosts
This will help avoid many socket timeout especially when we use
python-memcached with eventlet.

For example,When use eventlet,for each greenlet we call client.get
first time,the client will be rebuild,and client.servers[*].deaduntil
will be 0.For each http request,we will have one greenlet.So we get
a socket timeout for each request.
This could be fixed in application,but i believe it's better to fix
it in python-memcached.