Allow keys to be encoded before use.
Ported patch in #52 from @harlowja to current branch.  Added tests.

For the cases where the user wants to transparently
encode keys (say using urllib) before they are used
further allow a encoding function to be passed in that
will perform these types of activities (by default it
is the identity function).